+++ airflow/sensors/base_sensor_operator.py
@@ -18,6 +18,9 @@
 
 import hashlib
 from datetime import timedelta
+import inspect
+import os
+from functools import wraps
 from time import sleep
 from typing import Any, Dict, Iterable
 
@@ -176,3 +179,59 @@ class BaseSensorOperator(BaseOperator, SkipMixin):
         if self.reschedule:
             return BaseOperator.deps.fget(self) | {ReadyToRescheduleDep()}
         return BaseOperator.deps.fget(self)
+
+def poke_mode_only(cls):
+    """
+    Class Decorator for child classes of BaseSensorOperator to indicate
+    that instances of this class are only safe to use poke mode.
+
+    Will decorate all methods in the class to assert they did not change
+    the mode from 'poke'.
+
+    :param cls: BaseSensor class to enforce methods only use 'poke' mode.
+    :type cls: type
+    """
+    def decorate(cls):
+        if not issubclass(cls, BaseSensorOperator):
+            raise ValueError("poke_mode_only decorator should only be "
+                             + "applied to subclasses of BaseSensorOperator,"
+                             + f" got:{cls}.")
+        for method in inspect.getmembers(cls, inspect.isfunction):
+            setattr(cls, method[0], _poke_mode_only_func_decorator(method))
+        return cls
+
+    return decorate(cls)
+
+
+def _poke_mode_only_func_decorator(method):
+    """
+    Decorator that raises an error if method changed mode to anything but
+    'poke'.function
+    :param method: BaseSensor class to enforce methods only use 'poke' mode.
+    :type method: function
+    """
+    func = method[1]  # function object from inspect.get_attr
+
+    @wraps(func)
+    def wrapper(*args, **kwargs):
+        result = func(*args, **kwargs)
+        # Poor mans way of getting instnace w/o bound methods in python3
+        # checking if fist parameter is self (aka non-static method).
+        if isinstance(args[0], BaseSensorOperator):
+            instance = args[0]
+            try:
+                if instance.mode != 'poke':
+                    raise ValueError(
+                        f"cannot set mode to 'poke'.")
+            except AttributeError:  # this happens on _set_context
+                pass
+
+        return result
+
+    return wrapper
+
+
+if 'BUILDING_AIRFLOW_DOCS' in os.environ:
+    # flake8: noqa: F811
+    # Monkey patch hook to get good function headers while building docs
+    apply_defaults = lambda x: x
